renaming constants file
