+++ cea/optimization/slave/seasonal_storage/storage_main.py
@@ -17,7 +17,7 @@ save_file = 1
 
 import numpy as np
 import cea.optimization.slave.seasonal_storage.design_operation as StDesOp
-from cea.global_constants import HEAT_CAPACITY_OF_WATER_JPERKGK, DENSITY_OF_WATER_AT_60_DEGREES_KGPERM3, WH_TO_J
+from cea.constants import HEAT_CAPACITY_OF_WATER_JPERKGK, DENSITY_OF_WATER_AT_60_DEGREES_KGPERM3, WH_TO_J
 
 __author__ = "Tim Vollrath"
 __copyright__ = "Copyright 2015, Architecture and Building Systems - ETH Zurich"
